requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
google-generativeai==0.5.4
cloudinary==1.40.0
python-dotenv==1.0.1
cffi>=1.16.0
